Introduce pandas dataframes
Pandas builds on numpy to create high-performance query and
manipulation, which seems to fit the ecs paradigm rather well.
Introduced a component container abstraction so that we can use
data frames. This now works in a basic test, but it actually implies
our current system implementation is too high-level yet; we should
have an ever higher-level system abstraction that just gets entire
collections of components and does its own query of relevant ones.